Rename package and update version
Rename package and update version
