--- file_automation/__init__.py
@@ -1,31 +1,0 @@
-from file_automation.local.dir.dir_process import copy_dir, rename_dir, create_dir, remove_dir_tree
-from file_automation.local.file.file_process import copy_file, remove_file, rename_file, copy_specify_extension_file, \
-    copy_all_file_to_dir
-from file_automation.local.zip.zip_process import zip_dir, zip_file, zip_info, zip_file_info, set_zip_password, \
-    read_zip_file, unzip_file, unzip_all
-from file_automation.remote.google_drive.delete.delete_manager import drive_delete_file
-from file_automation.remote.google_drive.dir.folder_manager import drive_add_folder
-from file_automation.remote.google_drive.download.download_file import drive_download_file, \
-    drive_download_file_from_folder
-from file_automation.remote.google_drive.driver_instance import driver_instance
-from file_automation.remote.google_drive.search.search_drive import \
-    drive_search_all_file, drive_search_field, drive_search_file_mimetype
-from file_automation.remote.google_drive.share.share_file import \
-    drive_share_file_to_anyone, drive_share_file_to_domain, drive_share_file_to_user
-from file_automation.remote.google_drive.upload.upload_to_driver import \
-    drive_upload_dir_to_folder, drive_upload_to_folder, drive_upload_dir_to_drive, drive_upload_to_drive
-from file_automation.utils.executor.action_executor import execute_action, execute_files, add_command_to_executor
-from file_automation.utils.file_process.get_dir_file_list import get_dir_files_as_list
-from file_automation.utils.json.json_file import read_action_json
-from file_automation.utils.project.create_project_structure import create_project_dir
-
-__all__ = [
-    "copy_file", "rename_file", "remove_file", "copy_all_file_to_dir", "copy_specify_extension_file",
-    "copy_dir", "create_dir", "remove_dir_tree", "zip_dir", "zip_file", "zip_info",
-    "zip_file_info", "set_zip_password", "unzip_file", "read_zip_file",
-    "unzip_all", "driver_instance", "drive_search_all_file", "drive_search_field", "drive_search_file_mimetype",
-    "drive_upload_dir_to_folder", "drive_upload_to_folder", "drive_upload_dir_to_drive", "drive_upload_to_drive",
-    "drive_add_folder", "drive_share_file_to_anyone", "drive_share_file_to_domain", "drive_share_file_to_user",
-    "drive_delete_file", "drive_download_file", "drive_download_file_from_folder", "execute_action", "execute_files",
-    "add_command_to_executor", "read_action_json", "get_dir_files_as_list", "create_project_dir"
-]
