+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "automation_file"
-version = "0.0.16"
+version = "0.0.17"
 authors = [
     { name = "JE-Chen", email = "zenmailman@gmail.com" },
 ]
