+++ tests/unit_test/remote/google_drive/quick_test.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-from file_automation import driver_instance
-from file_automation.remote.google_drive.search.search_drive import drive_search_all_file
+from automation_file import driver_instance
+from automation_file.remote.google_drive.search.search_drive import drive_search_all_file
 
 driver_instance.later_init(str(Path(Path.cwd(), "token.json")), str(Path(Path.cwd(), "credentials.json")))
 print(drive_search_all_file())
